pyproject.toml
[project]
name = "fortress"
version = "0.1.0"
description = "Enterprise-grade User Management Service built with Django"
readme = "README.md"
requires-python = ">=3.11"

authors = [
    { name = "Akshat Tamrakar", email = "akshat1997tamrakar@gmail.com" }
]

dependencies = [
    "boto3>=1.42.16",
    "django>=5.0,<6.0",
    "django-redis>=5.4.0",
    "djangorestframework>=3.15",
    "gunicorn>=21.2.0",
    "pyjwt[crypto]>=2.8.0",
    "python-dotenv>=1.0.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6",
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-cov>=4.1.0",
    "hypothesis>=6.100.0",
]

# --------------------
# uv configuration
# --------------------
[tool.uv]
managed = true


# --------------------
# setuptools
# --------------------
[tool.setuptools.packages.find]
include = ["apps*","fortress*"]


# --------------------
# Ruff (lint + format)
# --------------------
[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

# --------------------
# Pytest
# --------------------
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "fortress.settings"
python_files = ["test_*.py", "*_test.py", "tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["apps"]
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
    "--cov=apps",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-branch",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "property: Property-based tests using Hypothesis",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["apps"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]